breaks up test-helper module
