+++ tests/helpers/bunch.py
@@ -1,0 +1,24 @@
+
+class Bunch:
+    """
+    Useful for mocking class instances.
+
+    In python, you cannot access dictionary keys with .[prop] notation,
+    e.g. you can access a property like this `mydict['myprop']`
+    but not like this `mydict.prop`
+
+    So when you want to mock an object that has properties,
+    you can't just use a dictionary. You *can* instead
+    just use Bunches like this:
+
+    ```
+    myObj = Bunch(
+        myProp = Bunch(myNestedProp = 'a')
+    )
+
+    print(myObj.myProp.myNestedProp) # prints 'a'
+    ```
+    """
+
+    def __init__(self, **kwds):
+        self.__dict__.update(kwds)
